+++ src/main/java/com/bizcub/randomItemSpeedrun/client/gui/GameStartScreen.java
@@ -42,10 +42,10 @@ public class GameStartScreen extends Screen {
 
         //? >=1.20.2 {
         WidgetSprites sprites = new WidgetSprites(Utils.getIdentifier("default"), Utils.getIdentifier("disabled"), Utils.getIdentifier("hovered"));
-        ImageButton settingsButton = this.addRenderableWidget(new ImageButton(getWidthPercent(1), getHeightPercent(2.5), getWidthPercent(3.8), getWidthPercent(3.8), sprites, button -> this.minecraft.setScreen(Compat.getScreen(this))));
+        ImageButton settingsButton = this.addRenderableWidget(new ImageButton(getWidthPercent(1), getHeightPercent(2.5), getWidthPercent(3.8), getWidthPercent(3.8), sprites, button -> this.minecraft.gui.setScreen(Compat.getScreen(this))));
         //?} else {
         /*var sprite = Utils.getIdentifier("textures/gui/sprites/widgets.png");
-        ImageButton settingsButton = this.addRenderableWidget(new ImageButton(getWidthPercent(1), getHeightPercent(2.5), getWidthPercent(3.8), getWidthPercent(3.8), 0, 0, getWidthPercent(3.9), sprite, getWidthPercent(12), getWidthPercent(12.1), button -> this.minecraft.setScreen(Compat.getScreen(this))));
+        ImageButton settingsButton = this.addRenderableWidget(new ImageButton(getWidthPercent(1), getHeightPercent(2.5), getWidthPercent(3.8), getWidthPercent(3.8), 0, 0, getWidthPercent(3.9), sprite, getWidthPercent(12), getWidthPercent(12.1), button -> this.minecraft.gui.setScreen(Compat.getScreen(this))));
         settingsButton.setPosition(getWidthPercent(1), getHeightPercent(2.5));*///?}
 
         addRenderableWidget(Button.builder(CommonComponents.GUI_DONE, button -> onClose()).pos(this.width / 2 + 4, getHeightPercent(91.5)).size(125, 20).build());